refactor: image validation 서비스 분리
