+++ backend/src/main/java/team/teamby/teambyteam/filesystem/exception/NotAllowedImageExtensionException.java
@@ -1,9 +1,9 @@
-package team.teamby.teambyteam.feed.exception;
+package team.teamby.teambyteam.filesystem.exception;
 
 import team.teamby.teambyteam.common.exception.CustomBadRequestException;
 
-public class FeedNotAllowedImageExtensionException extends CustomBadRequestException {
-    public FeedNotAllowedImageExtensionException(final String imageName) {
+public class NotAllowedImageExtensionException extends CustomBadRequestException {
+    public NotAllowedImageExtensionException(final String imageName) {
         super(String.format("허용되지 않은 확장자입니다. - request info { image_name : %s }", imageName)
         );
     }
