+++ backend/src/test/java/team/teamby/teambyteam/feed/application/FeedWriteServiceTest.java
@@ -11,10 +11,10 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.web.multipart.MultipartFile;
 import team.teamby.teambyteam.common.ServiceTest;
 import team.teamby.teambyteam.feed.application.dto.FeedThreadWritingRequest;
-import team.teamby.teambyteam.feed.exception.FeedImageSizeException;
-import team.teamby.teambyteam.feed.exception.FeedNotAllowedImageExtensionException;
 import team.teamby.teambyteam.feed.exception.FeedWritingRequestEmptyException;
 import team.teamby.teambyteam.filesystem.FileStorageManager;
+import team.teamby.teambyteam.filesystem.exception.ImageSizeException;
+import team.teamby.teambyteam.filesystem.exception.NotAllowedImageExtensionException;
 import team.teamby.teambyteam.member.configuration.dto.MemberEmailDto;
 import team.teamby.teambyteam.member.domain.Member;
 import team.teamby.teambyteam.member.exception.MemberNotFoundException;
@@ -89,7 +89,7 @@ class FeedWriteServiceTest extends ServiceTest {
             // when & then
             assertThatThrownBy(() -> feedWriteService.write(request, new MemberEmailDto(author.getEmail().getValue()),
                     teamPlace.getId()))
-                    .isInstanceOf(FeedImageSizeException.class)
+                    .isInstanceOf(ImageSizeException.class)
                     .hasMessageContaining("허용된 이미지의 크기를 초과했습니다.");
         }
 
@@ -104,7 +104,7 @@ class FeedWriteServiceTest extends ServiceTest {
             // when & then
             assertThatThrownBy(() -> feedWriteService.write(request, new MemberEmailDto(author.getEmail().getValue()),
                     teamPlace.getId()))
-                    .isInstanceOf(FeedNotAllowedImageExtensionException.class)
+                    .isInstanceOf(NotAllowedImageExtensionException.class)
                     .hasMessageContaining("허용되지 않은 확장자입니다.");
         }
